DDF-1323: conformed more to the decorator pattern.
