+++ platform/admin/core/admin-core-api/src/test/java/org/codice/ddf/ui/admin/api/module/DecoratorTest.java
@@ -27,7 +27,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mockito;
 
-public class AdminModuleDecoratorTest {
+public class DecoratorTest {
 
     private AdminModule module;
 
@@ -40,37 +40,15 @@ public class AdminModuleDecoratorTest {
     public void testAdminModuleAdminModuleDecorator() {
         doReturn("test").when(module).getName();
         doReturn("0").when(module).getId();
-        AdminModuleDecorator proxy = new AdminModuleDecorator(module);
+        ValidationDecorator proxy = new ValidationDecorator(module);
         assertThat("values are proxied correctly", proxy.getName(), is(module.getName()));
         assertThat("values are proxied correctly", proxy.getId(), is(module.getId()));
     }
 
     @Test
-    public void testValidAdminModule() throws URISyntaxException {
-        doReturn(new URI("js/app.js")).when(module).getJSLocation();
-        AdminModuleDecorator proxy = new AdminModuleDecorator(module);
-        assertThat("relative paths are valid", proxy.isValid(), is(true));
-    }
-
-    @Test
-    public void testInvalidAdminModule() throws URISyntaxException {
-        doReturn(new URI("http://test/js/app.js")).when(module).getJSLocation();
-        AdminModuleDecorator proxy = new AdminModuleDecorator(module);
-        assertThat("absolute paths are not valid", proxy.isValid(), is(false));
-    }
-
-    @Test
-    public void testPartialInvalidAdminModule() throws URISyntaxException {
-        doReturn(new URI("js/app.js")).when(module).getJSLocation();
-        doReturn(new URI("/css/styles.css")).when(module).getCSSLocation();
-        AdminModuleDecorator proxy = new AdminModuleDecorator(module);
-        assertThat("any absolute paths are not valid", proxy.isValid(), is(false));
-    }
-
-    @Test
     public void testToMap() {
         doReturn("test").when(module).getName();
-        AdminModuleDecorator proxy = new AdminModuleDecorator(module);
+        Decorator proxy = new ValidationDecorator(module);
         String name = (String) proxy.toMap().get("name");
         assertThat("hash map gets constructed correctly", name, is("test"));
     }
@@ -79,14 +57,14 @@ public class AdminModuleDecoratorTest {
     public void testURLToMap() throws URISyntaxException {
         String uri = "js/app.js";
         doReturn(new URI(uri)).when(module).getJSLocation();
-        AdminModuleDecorator proxy = new AdminModuleDecorator(module);
+        Decorator proxy = new ValidationDecorator(module);
         String jsLocation = (String) proxy.toMap().get("jsLocation");
         assertThat("URLs get mapped to strings", jsLocation, is(uri));
     }
 
     @Test
     public void testNullURLToMap() {
-        AdminModuleDecorator proxy = new AdminModuleDecorator(module);
+        Decorator proxy = new ValidationDecorator(module);
         String jsLocation = (String) proxy.toMap().get("jsLocation");
         assertThat("null URLs get mapped to empty strings", jsLocation, is(""));
     }
@@ -94,7 +72,7 @@ public class AdminModuleDecoratorTest {
     @Test
     public void testCompareTo() {
         doReturn("test").when(module).getName();
-        AdminModuleDecorator proxy = new AdminModuleDecorator(module);
+        Decorator proxy = new ValidationDecorator(module);
         assertThat("modules are the same if the have the same name", proxy.compareTo(module),
                 is(0));
     }
@@ -110,7 +88,7 @@ public class AdminModuleDecoratorTest {
         list.add(world);
         list.add(hello);
 
-        List<AdminModuleDecorator> modules = AdminModuleDecorator.wrap(list);
+        List<ValidationDecorator> modules = ValidationDecorator.wrap(list);
         Collections.sort(modules);
 
         assertThat("modules are sorted lexographically by name", modules.get(0).getName(),
